Updated python files

+++ SE Raspberry/Webcam/New/captures.py
@@ -1,11 +1,13 @@
 import os
 
+'''
 def set_affinity(core_id):
     command = f"taskset -cp {core_id} {os.getpid()}"
     os.system(command)
 
 # Set the current process to run on core 0
 set_affinity(0)
+'''
 
 import time
 import cv2
@@ -43,25 +45,26 @@ def capture_images():
         os.makedirs(save_dir)
 
     # Define the size of the shared memory block (RGB image size)
-    shm_size = 800 * 600 * 3
+    shm_size = 1640 * 1232 * 3
 
     # Open the existing shared memory block
     shm = shared_memory.SharedMemory(name='shm_camera')
 
-    first_captured = False
 
     try:
         while True:
-            image_data = np.ndarray(shape=(600, 800, 3), dtype=np.uint8, buffer=shm.buf)
+            #print("capture")
+            image_data = np.ndarray(shape=(1232, 1640, 3), dtype=np.uint8, buffer=shm.buf)
             
             compressed_image_bytes = compress_image(image_data)
             image_base64 = base64.b64encode(compressed_image_bytes).decode('utf-8')
             server_data = {"type": "camera image", "content": image_base64, "uuid": ""}
-            response = requests.post(url, json=server_data)
+            requests.post(url, json=server_data)
             #print(f"Response: {response.json()}")
             
             last_image_filename = os.path.join(save_dir, f'last_image.png')
-            cv2.imwrite(last_image_filename, image_data)
+            if not os.path.exists(last_image_filename):
+                cv2.imwrite(last_image_filename, image_data)
         
             
             time.sleep(0.14)
